+++ src/Comparison/Exception/VersionComparisonException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Version\Comparison\Exception;
+
+use Throwable;
+
+interface VersionComparisonException extends Throwable
+{
+}