+++ src/Exception/VersionException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Version\Exception;
 
-interface ExceptionInterface
+use Throwable;
+
+interface VersionException extends Throwable
 {
 }