+++ src/VersionsCollection.php
@@ -9,7 +9,6 @@ use IteratorAggregate;
 use ArrayIterator;
 use Traversable;
 use Version\Comparison\Constraint\Constraint;
-use Version\Exception\CollectionIsEmptyException;
 
 class VersionsCollection implements Countable, IteratorAggregate
 {
@@ -34,43 +33,19 @@ class VersionsCollection implements Countable, IteratorAggregate
         return empty($this->versions);
     }
 
-    public function first(): Version
+    public function first(): ?Version
     {
-        if (empty($this->versions)) {
-            throw new CollectionIsEmptyException('Invoking first() on an empty collection');
-        }
-
-        return $this->versions[0];
+        return $this->versions[0] ?? null;
     }
 
-    public function last(): Version
+    public function last(): ?Version
     {
-        if (empty($this->versions)) {
-            throw new CollectionIsEmptyException('Invoking last() on an empty collection');
-        }
-
-        return $this->versions[count($this->versions) - 1];
+        return $this->versions[count($this->versions) - 1] ?? null;
     }
 
     public function getIterator(): Traversable
     {
         return new ArrayIterator($this->versions);
-    }
-
-    /**
-     * @deprecated This method will be removed in 4.0.0. Use sorted() instead.
-     */
-    public function sort(string $direction = self::SORT_ASC): void
-    {
-        usort($this->versions, function (Version $a, Version $b) use ($direction) {
-            $result = $a->compareTo($b);
-
-            if ($direction === self::SORT_DESC) {
-                $result *= -1;
-            }
-
-            return $result;
-        });
     }
 
     public function sortedAscending(): VersionsCollection